+++ rye/src/cli/init.rs
@@ -35,6 +35,9 @@ pub struct Args {
     /// Where to place the project (defaults to current path)
     #[arg(default_value = ".")]
     path: PathBuf,
+    /// Initialization type
+    #[command(flatten)]
+    init_type: ArgTemplateChoice,
     /// Minimal Python version supported by this project.
     #[arg(long)]
     min_py: Option<String>,
@@ -82,174 +85,51 @@ pub struct Args {
     quiet: bool,
 }
 
+#[derive(Parser, Debug)]
+#[group(multiple = false)]
+struct ArgTemplateChoice {
+    /// Generate a library project (default).
+    #[arg(long)]
+    lib: bool,
+
+    /// Generate an executable project.
+    #[arg(long)]
+    script: bool,
+}
+
+enum TemplateChoice {
+    Lib,
+    Script,
+}
+
 /// The pyproject.toml template
-///
-/// This uses a template just to simplify the flexibility of emitting it.
-const TOML_TEMPLATE: &str = r#"[project]
-name = {{ name }}
-version = {{ version }}
-description = {{ description }}
-{%- if author %}
-authors = [
-    { name = {{ author[0] }}, email = {{ author[1] }} }
-]
-{%- endif %}
-{%- if dependencies %}
-dependencies = [
-{%- for dependency in dependencies %}
-    {{ dependency }},
-{%- endfor %}
-]
-{%- else %}
-dependencies = []
-{%- endif %}
-{%- if with_readme %}
-readme = "README.md"
-{%- endif %}
-requires-python = {{ requires_python }}
-{%- if license %}
-license = { text = {{ license }} }
-{%- endif %}
-{%- if private %}
-classifiers = ["Private :: Do Not Upload"]
-{%- endif %}
+const TOML_TEMPLATE: &str = include_str!("../templates/pyproject.toml.j2");
 
-[project.scripts]
-hello = {{ name_safe ~ ":hello"}}
+/// The template for the README.md.
+const README_TEMPLATE: &str = include_str!("../templates/README.md.j2");
 
-{%- if not is_virtual %}
+/// The template for the LICENSE.txt.
+const LICENSE_TEMPLATE: &str = include_str!("../templates/LICENSE.txt.j2");
 
-[build-system]
-{%- if build_system == "hatchling" %}
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-{%- elif build_system == "setuptools" %}
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
-{%- elif build_system == "flit" %}
-requires = ["flit_core>=3.4"]
-build-backend = "flit_core.buildapi"
-{%- elif build_system == "pdm" %}
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
-{%- elif build_system == "maturin" %}
-requires = ["maturin>=1.2,<2.0"]
-build-backend = "maturin"
-{%- endif %}
-{%- endif %}
+/// Template for the __init__.py when --lib is specified.
+const INIT_PY_LIB_TEMPLATE: &str = include_str!("../templates/lib/default/__init__.py.j2");
 
-[tool.rye]
-managed = true
-{%- if is_virtual %}
-virtual = true
-{%- endif %}
-{%- if dev_dependencies %}
-dev-dependencies = [
-{%- for dependency in dev_dependencies %}
-    {{ dependency }},
-{%- endfor %}
-]
-{%- else %}
-dev-dependencies = []
-{%- endif %}
+/// Template for the __init__.py when --script is specified.
+const INIT_PY_BIN_TEMPLATE: &str = include_str!("../templates/script/default/__init__.py.j2");
 
-{%- if not is_virtual %}
-{%- if build_system == "hatchling" %}
+const INIT_PY_BIN_MAIN_TEMPLATE: &str = include_str!("../templates/script/default/__main__.py.j2");
 
-[tool.hatch.metadata]
-allow-direct-references = true
+/// Template for the lib.rs when using the maturin build system.
+const LIB_RS_TEMPLATE: &str = include_str!("../templates/lib/maturin/lib.rs.j2");
 
-[tool.hatch.build.targets.wheel]
-packages = [{{ "src/" ~ name_safe }}]
-{%- elif build_system == "maturin" %}
+/// Template for the __init__.py when using the maturin build system.
+const RUST_INIT_PY_TEMPLATE: &str = include_str!("../templates/lib/maturin/__init__.py.j2");
 
-[tool.maturin]
-python-source = "python"
-module-name = {{ name_safe ~ "._lowlevel" }}
-features = ["pyo3/extension-module"]
-{%- endif %}
-{%- endif %}
+/// Template for the Cargo.toml.
+const CARGO_TOML_TEMPLATE: &str = include_str!("../templates/lib/maturin/Cargo.toml.j2");
 
-"#;
-
-/// The template for the readme file.
-const README_TEMPLATE: &str = r#"# {{ name }}
-
-Describe your project here.
-
-{%- if license %}
-* License: {{ license }}
-{%- endif %}
-
-"#;
-
-const LICENSE_TEMPLATE: &str = r#"
-{{ license_text }}
-"#;
-
-/// Template for the __init__.py
-const INIT_PY_TEMPLATE: &str = r#"def hello():
-    return "Hello from {{ name }}!"
-
-"#;
-
-/// Template for the lib.rs
-const LIB_RS_TEMPLATE: &str = r#"use pyo3::prelude::*;
-
-/// Prints a message.
-#[pyfunction]
-fn hello() -> PyResult<String> {
-    Ok("Hello from {{ name }}!".into())
-}
-
-/// A Python module implemented in Rust.
-#[pymodule]
-fn _lowlevel(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_pyfunction!(hello, m)?)?;
-    Ok(())
-}
-"#;
-
-/// Template for the __init__.py
-const RUST_INIT_PY_TEMPLATE: &str = r#"from {{ name_safe }}._lowlevel import hello
-__all__ = ["hello"]
-
-"#;
-
-/// Template for the Cargo.toml
-const CARGO_TOML_TEMPLATE: &str = r#"[package]
-name = {{ name }}
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-name = {{ name_safe }}
-crate-type = ["cdylib"]
-
-[dependencies]
-pyo3 = "0.19.0"
-
-"#;
-
-/// Template for fresh gitignore files
-const GITIGNORE_TEMPLATE: &str = r#"# python generated files
-__pycache__/
-*.py[oc]
-build/
-dist/
-wheels/
-*.egg-info
-
-{%- if is_rust %}
-# Rust
-target/
-{%- endif %}
-
-# venv
-.venv
-
-"#;
+/// Template for fresh gitignore files.
+const GITIGNORE_TEMPLATE: &str = include_str!("../templates/gitignore.j2");
 
 /// Script used for setup.py setup proxy.
 const SETUP_PY_PROXY_SCRIPT: &str = r#"
@@ -399,7 +279,7 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
         true
     } else if !cmd.no_readme {
         let rv = env.render_named_str(
-            "README.txt",
+            "README.md",
             README_TEMPLATE,
             context! {
                 name => metadata.name,
@@ -419,6 +299,20 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
 
     let private = cmd.private;
 
+    // What template are we using?
+    let template = {
+        if cmd.init_type.script {
+            TemplateChoice::Script
+        } else {
+            // default value
+            TemplateChoice::Lib
+        }
+    };
+
+    if cmd.init_type.script && build_system == BuildSystem::Maturin {
+        bail!("--script is not supported when the build-system is maturin");
+    }
+
     // crate a python module safe name.  This is the name on the metadata with
     // underscores instead of dashes to form a valid python package name and in
     // case it starts with a digit, an underscore is prepended.
@@ -430,8 +324,6 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
     {
         name_safe.insert(0, '_');
     }
-
-    let is_rust = build_system == BuildSystem::Maturin;
 
     // if git init is successful prepare the local git repository
     if !is_inside_git_work_tree(&dir)
@@ -452,7 +344,7 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
                 "gitignore.txt",
                 GITIGNORE_TEMPLATE,
                 context! {
-                    is_rust
+                    is_rust => matches!(build_system, BuildSystem::Maturin)
                 },
             )?;
             fs::write(&gitignore, rv).context("failed to write .gitignore")?;
@@ -478,6 +370,7 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
             license => metadata.license,
             dependencies => metadata.dependencies,
             dev_dependencies => metadata.dev_dependencies,
+            is_script => matches!(template, TemplateChoice::Script),
             is_virtual,
             with_readme,
             build_system,
@@ -490,37 +383,63 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
         let src_dir = dir.join("src");
         if !imported_something && !src_dir.is_dir() {
             let name = metadata.name.expect("project name");
-            if is_rust {
-                fs::create_dir_all(&src_dir).ok();
-                let project_dir = dir.join("python").join(&name_safe);
-                fs::create_dir_all(&project_dir).ok();
-                let rv = env.render_named_str("lib.rs", LIB_RS_TEMPLATE, context! { name })?;
-                fs::write(src_dir.join("lib.rs"), rv).context("failed to write lib.rs")?;
-                let rv = env.render_named_str(
-                    "Cargo.json",
-                    CARGO_TOML_TEMPLATE,
-                    context! {
-                        name,
-                        name_safe,
-                    },
-                )?;
-                fs::write(dir.join("Cargo.toml"), rv).context("failed to write Cargo.toml")?;
-                let rv = env.render_named_str(
-                    "__init__.py",
-                    RUST_INIT_PY_TEMPLATE,
-                    context! {
-                        name_safe
-                    },
-                )?;
-                fs::write(project_dir.join("__init__.py"), rv)
-                    .context("failed to write __init__.py")?;
-            } else {
-                let project_dir = src_dir.join(&name_safe);
-                fs::create_dir_all(&project_dir).ok();
-                let rv =
-                    env.render_named_str("__init__.py", INIT_PY_TEMPLATE, context! { name })?;
-                fs::write(project_dir.join("__init__.py"), rv)
-                    .context("failed to write __init__.py")?;
+            match (template, build_system) {
+                (TemplateChoice::Lib, BuildSystem::Maturin) => {
+                    fs::create_dir_all(&src_dir).ok();
+                    let project_dir = dir.join("python").join(&name_safe);
+                    fs::create_dir_all(&project_dir).ok();
+                    let rv = env.render_named_str("lib.rs", LIB_RS_TEMPLATE, context! { name })?;
+                    fs::write(src_dir.join("lib.rs"), rv).context("failed to write lib.rs")?;
+                    let rv = env.render_named_str(
+                        "Cargo.json",
+                        CARGO_TOML_TEMPLATE,
+                        context! {
+                            name,
+                            name_safe,
+                        },
+                    )?;
+                    fs::write(dir.join("Cargo.toml"), rv).context("failed to write Cargo.toml")?;
+                    let rv = env.render_named_str(
+                        "__init__.py",
+                        RUST_INIT_PY_TEMPLATE,
+                        context! {
+                            name_safe
+                        },
+                    )?;
+                    fs::write(project_dir.join("__init__.py"), rv)
+                        .context("failed to write __init__.py")?;
+                }
+                (TemplateChoice::Lib, _) => {
+                    let project_dir = src_dir.join(&name_safe);
+                    fs::create_dir_all(&project_dir).ok();
+                    let rv = env.render_named_str(
+                        "__init__.py",
+                        INIT_PY_LIB_TEMPLATE,
+                        context! { name },
+                    )?;
+                    fs::write(project_dir.join("__init__.py"), rv)
+                        .context("failed to write __init__.py")?;
+                }
+                (TemplateChoice::Script, _) => {
+                    let project_dir = src_dir.join(&name_safe);
+                    fs::create_dir_all(&project_dir).ok();
+
+                    let rv1 = env.render_named_str(
+                        "__init__.py",
+                        INIT_PY_BIN_TEMPLATE,
+                        context! { name },
+                    )?;
+                    fs::write(project_dir.join("__init__.py"), rv1)
+                        .context("failed to write __init__.py")?;
+
+                    let rv2 = env.render_named_str(
+                        "__main__.py",
+                        INIT_PY_BIN_MAIN_TEMPLATE,
+                        context! { name_safe },
+                    )?;
+                    fs::write(project_dir.join("__main__.py"), rv2)
+                        .context("failed to write __main__.py")?;
+                }
             }
         }
     }